friends_accounts_project/pyproject.toml
[tool.poetry]
name = "django-api-friends-accounts"
version = "0.1.0"
description = ""
authors = ["back20 <ivankovavm.new@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
